upcoming

+++ myproject/blog/views/home_view.py
@@ -1,11 +1,14 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from ..models import Blog
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 # Create your views here.
 
 
 
 # Create a new blog post
+@login_required
 def create_blog(request):
     if request.method == 'POST':
         title = request.POST.get('title')
@@ -13,7 +16,7 @@ def create_blog(request):
         description = request.POST.get('description')
         image = request.FILES.get('image')
 
-        blog = Blog(title=title, subtitle=subtitle, description=description, image=image)
+        blog = Blog(title=title, subtitle=subtitle, description=description, image=image, author=request.user)
         blog.save()
 
         return redirect('blog_list')  # Redirect to the list of blog posts
@@ -51,7 +54,13 @@ def update_blog(request, blog_id):
 def delete_blog(request, blog_id):
     blog = get_object_or_404(Blog, pk=blog_id)
     if request.method == 'POST':
-        blog.delete()
-        return redirect('blog_list')
+        # Check if the logged-in user is the author
+        if blog.author == request.user:
+            blog.delete()
+            messages.success(request, "Blog post deleted successfully.")
+            return redirect('blog_list')
+        else:
+            messages.error(request, "You are not authorized to delete this blog post.")
+            return redirect('blog_detail', blog_id=blog.id)
 
     return render(request, 'home/delete_blog.html', {'blog': blog})
